+++ lib/state_manager/state_manager.dart
@@ -1,0 +1,15 @@
+///
+/// Created by
+///
+/// ─▄▀─▄▀
+/// ──▀──▀
+/// █▀▀▀▀▀█▄
+/// █░░░░░█─█
+/// ▀▄▄▄▄▄▀▀
+///
+/// Rafaelbarbosatec
+/// on 23/02/22
+export 'bonfire_injector.dart';
+export 'state_controller.dart';
+export 'state_controller_consumer.dart';
+export 'use_state_controller.dart';